+++ gocoinlib/src/impl/Client.class.php
@@ -239,28 +239,25 @@ class Client
    */
   public function getError($as_is=FALSE)
   {
+    $error = $this -> error;
     //support returning the error as-is
-    if ($as_is) { return $this -> error; }
+    if ($as_is) { return $error; }
 
     //otherwise, potentially translate it
-    if (is_object($this -> error) && get_class($this -> error) == 'stdClass')
+    if (is_object($error) && get_class($error) == 'stdClass')
     {
-      $errors = array();
-      foreach (get_object_vars($this -> error) as $key => $value)
-      {
-        if (is_array($value))
-        {
-          $errors[$key] = implode(',',$value);
-        }
-        else
-        {
-          $errors[$key] = $value;
+      if (isset($error -> error)) {
+       return $error -> error;
+      }
+      $err = $error->message . ' ';
+      if (isset($error -> errors)) {
+        foreach ($error -> errors as $key => $value) {
+          $err .= $key . ' ' . $value[0] . ' ';
         }
       }
-      $err = implode(',',array_values($errors));
-      return $err;
+      return trim($err);
     }
-    return $this -> error;
+    return $error;
   }
 
   /**
@@ -458,15 +455,9 @@ class Client
     $result = json_decode($result);
 
     //make sure there isn't an error
-    if (isset($result -> error))
+    if (isset($result->status) && (intval($result->status) >= 400))
     {
-      $this -> setError($result -> error_description);
-      return FALSE;
-    }
-    //make sure there isn't an error
-    if (isset($result -> errors))
-    {
-      $this -> setError($result -> errors);
+      $this -> setError($result);
       return FALSE;
     }